refactor(scripts): decompose single-story-close into phases/
Extracts the eight close-pipeline phases plus their private helpers
from the monolithic single-story-close.js (866 LOC) into a sibling
phases/ directory under .agents/scripts/lib/orchestration/single-story-close/:

  - base-sync.js          (194 LOC) — origin/<base> sync + friction-comment
                                       on conflict + agent::blocked flip
  - code-review.js        (187 LOC) — story-scope review + cross-ref comment
  - auto-merge.js         (123 LOC) — gh pr merge --auto --squash
  - pull-request.js       (114 LOC) — gh pr list probe + gh pr create
  - close-validation.js   ( 79 LOC) — canonical close-validation gate chain
  - worktree-reap.js      ( 73 LOC) — worktree reap + trace-hook env clear
  - options.js            ( 70 LOC) — CLI / injection option parsing
  - push.js               ( 42 LOC) — git push -u origin <storyBranch>

The parent single-story-close.js shrinks to 292 LOC and becomes a thin
CLI entry that sequences the phases via dependency injection. The
parent's existing test suite mocks IO primitives by module URL with
?t= cache-busting — naive static imports in the extracted phases would
bind to the un-mocked symbols since the phase modules are cached once.
The decomposition therefore threads gitSync, execFileSync,
WorktreeManager, runCloseValidation, and buildGatesFromConfig as
injection params from the parent SUT, plus retains thin wrappers
(ensurePullRequest / enableAutoMerge) for tests that call those names
directly.

Pure code-move refactor; no behavior change. gh-spawn call sites kept
verbatim (separate concern — Story #2990 handles routing them through
providers/github/). 6751/6751 tests pass.

Refs #3003 #2988
